# backend/src/swagger-complete.yaml
openapi: 3.0.0
info:
  title: Jarvis Meeting Assistant API
  description: Complete API for meeting transcription, contact management, calendar sync, and automation
  version: 1.0.0
  contact:
    name: Jarvis Support
    email: support@jarvis.com
servers:
  - url: http://localhost:8080/api
    description: Local development server
  - url: https://your-production-url.com/api
    description: Production server
security:
  - UserAuth: []
components:
  securitySchemes:
    UserAuth:
      type: apiKey
      in: header
      name: x-user-id
      description: User ID for authentication
  schemas:
...
    Meeting:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        date:
          type: string
        duration:
          type: string
        transcript:
          type: string
        summary:
          type: string
        keyPoints:
          type: array
          items:
            type: string
        decisions:
          type: array
          items:
            type: string
        actionItems:
          type: array
          items:
            type: object
        originalFilename:
          type: string
        fileHash:
          type: string
        fileSizeMB:
          type: string
...
    ProcessingJob:
      type: object
      properties:
        jobId:
          type: string
        status:
          type: string
          enum: [processing, completed, failed]
        progress:
          type: integer
        meetingId:
          type: string
        error:
          type: string
paths:
...
  /meetings/upload:
    post:
      summary: Upload and process meeting audio
      description: Upload audio file with optional metadata. Returns job ID for status tracking.
      tags: [Meetings]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - audio
              properties:
                audio:
                  type: string
                  format: binary
                  description: Audio file (m4a, mp3, wav, webm)
                title:
                  type: string
                  description: Meeting title
                  example: "Q4 Planning Meeting"
                projectId:
                  type: string
                  description: Project ID
                  example: "proj-123"
                attendeeIds:
                  type: string
                  description: JSON array of contact IDs
                  example: '["contact-1", "contact-2"]'
      responses:
        '202':
          description: Upload accepted, processing in background
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  jobId:
                    type: string
                  status:
                    type: string
                  estimatedTimeSeconds:
                    type: integer
        '400':
          description: Bad request
        '401':
          description: Unauthorized
  /meetings/status/{jobId}:
    get:
      summary: Check processing job status
      tags: [Meetings]
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
          description: Job ID from upload response
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingJob'
  /meetings:
    get:
      summary: Get all meetings
      tags: [Meetings]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
          description: Maximum number of meetings
      responses:
        '200':
          description: List of meetings
          content:
            application/json:
              schema:
                type: object
                properties:
                  meetings:
                    type: array
                    items:
                      $ref: '#/components/schemas/Meeting'
                  count:
                    type: integer
  /meetings/{id}:
    get:
      summary: Get specific meeting
      tags: [Meetings]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Meeting details
          content:
            application/json:
              schema:
                type: object
                properties:
                  meeting:
                    $ref: '#/components/schemas/Meeting'
    delete:
      summary: Delete meeting
      tags: [Meetings]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Meeting deleted
  /meetings/{id}/generate-pdf:
    post:
      summary: Generate PDF for meeting
      tags: [Meetings]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: PDF generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  pdfUrl:
                    type: string
  /meetings/{id}/email:
    post:
      summary: Email meeting minutes
      tags: [Meetings]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Email sent
  /meetings/retry/{jobId}:
    post:
      summary: Retry failed job
      tags: [Meetings]
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Retry initiated
        '501':
          description: Feature not implemented

  # ==================== CONTACTS ====================
...